requirements.txt
matplotlib==3.4.3
numpy==1.20.2
pandas==1.2.4
python-dotenv==1.0.0
